chore: refactor API views into separate modules/files
